+++ erp-modules/erp-system/src/main/java/com/erp/system/controller/SysDictTypeController.java
@@ -1,8 +1,10 @@
 package com.erp.system.controller;
 
 import com.erp.common.core.domain.R;
+import com.erp.common.core.domain.ResultCode;
 import com.erp.system.domain.SysDictType;
 import com.erp.system.service.ISysDictTypeService;
+import com.erp.system.support.TenantWriteGuard;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -39,6 +41,9 @@ public class SysDictTypeController {
     @PreAuthorize("@ss.hasPermi('system:dict:add')")
     @PostMapping
     public R<Void> add(@RequestBody SysDictType dictType) {
+        if (!TenantWriteGuard.canWriteGlobalData()) {
+            return R.failed(ResultCode.FORBIDDEN);
+        }
         dictTypeService.save(dictType);
         return R.success();
     }
@@ -47,6 +52,9 @@ public class SysDictTypeController {
     @PreAuthorize("@ss.hasPermi('system:dict:edit')")
     @PutMapping
     public R<Void> edit(@RequestBody SysDictType dictType) {
+        if (!TenantWriteGuard.canWriteGlobalData()) {
+            return R.failed(ResultCode.FORBIDDEN);
+        }
         dictTypeService.updateById(dictType);
         return R.success();
     }
@@ -55,6 +63,9 @@ public class SysDictTypeController {
     @PreAuthorize("@ss.hasPermi('system:dict:remove')")
     @DeleteMapping("/{dictIds}")
     public R<Void> remove(@PathVariable("dictIds") List<Long> dictIds) {
+        if (!TenantWriteGuard.canWriteGlobalData()) {
+            return R.failed(ResultCode.FORBIDDEN);
+        }
         dictTypeService.removeByIds(dictIds);
         return R.success();
     }
